requirements.txt
asgiref==3.2.3
CodeConvert==3.0.2
colorama==0.4.3
Django==2.2.3
django-six==1.0.4
djangorestframework==3.11.0
entrypoints==0.3
filelock==3.0.12
flake8==3.7.9
importlib-metadata==1.3.0
mccabe==0.6.1
more-itertools==8.0.2
packaging==20.0
pluggy==0.13.1
py==1.8.1
pycodestyle==2.5.0
pyflakes==2.1.1
pyparsing==2.4.6
pytz==2019.3
short-url==1.2.2
shortuuid==0.5.0
six==1.13.0
sqlparse==0.3.0
toml==0.10.0
tox==3.14.3
virtualenv==16.7.9
zipp==0.6.0
